Separate common functions for sink and source
